Add files via upload
In this section we have analysed an array of Axially & transversely coupled diamond chain with N=1 where N is the number of atomic sites in each arm of the diamond. Now the on-site potential is set to be zero (close to zero for numerical stability) and we found extreme compact localization at half flux quantum . This result is important as according to P.W.Anderson , if there is disorder in system there is definitely be a localization, termed as localization, but in our case we have taken a completely periodic ordered structure. The localization character via flux tunability is extremely important as it can be used to fabricate different kind of spin filters & optoelectronic devices. 
Also this structure can be used to trap photons for a certain amount of time (Sebabrata Mukherjee et. al.).

Our research report is also uploaded for further clarification.